% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibbleColumns.R
\name{bind_rows_keyword}
\alias{bind_rows_keyword}
\title{Bind data frames together by row based on a keyword}
\usage{
bind_rows_keyword(df, keyword, env = NULL)
}
\arguments{
\item{df, keyword, env}{dataframe to take in current state of tibble keyword string and environment object to search in}
}
\description{
Allows user to pass a keyword and an environment to bind all data frames in that
environment with that keyword in the name by rows.
}
\examples{
mtcars \%>\% tbl_df() \%>\% tbl_module(filter(.,hp < 00), "c1") \%>\% filter(hp > 200) \%>\% 
tbl_out("c") \%>\% bind_rows_keyword("c")
}
