% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibbleColumns.R
\name{bool_to_binary}
\alias{bool_to_binary}
\title{New binary numeric column mirroring logical column}
\usage{
bool_to_binary(df, col)
}
\arguments{
\item{df, col}{a data frame and logical TRUE FALSE colum to change to binary}
}
\description{
Look at a logi cal TRUE FALSE column and make new column with binary represenation
}
\examples{
iris \%>\% mutate(Setosa = str_detect(.$Species, "setosa")) \%>\% tbl_bool_to_binary(.,Setosa)
}
