% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibbleColumns.R
\name{change_XoX_column_group}
\alias{change_XoX_column_group}
\title{General X over X Change Column by Group}
\usage{
change_XoX_column_group(df, col1, col2, XoX)
}
\description{
Creates a change column based on a group. This function is specific 
as the data must have three columns at most. A category group column 
group a/b, device type, segment, a calendar group month, year, day 
and a numeric column to aggregate users, visits, clicks etc.. The 
data columns MUST be in that order as well. Category Group, Calendar 
Group, Numeric Aggregate.
}
\examples{
tb \%>\% select(Type, Month, Users) \%>\% change_XoX_column_group(Dec,Jan,"MoM")
}
