% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibbleColumns.R
\name{file_choose}
\alias{file_choose}
\title{Tidy wrapper for file.choose function with doc format options}
\usage{
file_choose(type, sheet = NULL)
}
\arguments{
\item{type}{format type for read in document either csv or xls}
}
\description{
Opens GUI to select specific file and read in as either csv or xls with a sheet option for xls file
}
\examples{
file_choose("csv")
}
