% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibbleColumns.R
\name{lag_col}
\alias{lag_col}
\title{Created lag column based on anothwer}
\usage{
lag_col(df, col, replace_na_with = NULL)
}
\arguments{
\item{df, col, replace_na_with}{a data frame a column and what to replace NAs with}
}
\description{
Select a column to have the lag row of be noted in a new mutated column
}
\examples{
mtcars \%>\% lead_col(cyl,0)
}
