% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibbleColumns.R
\name{lm_summary_tibble}
\alias{lm_summary_tibble}
\title{Linear Model Summary Tibble}
\usage{
lm_summary_tibble(df, dep)
}
\arguments{
\item{df, dep}{a dataframe and a dependent variable name}
}
\description{
This function returns a tidy tibble output of the most important parts
of a lm summary a la the broom package.
}
\examples{
mtcars \%>\% select(mpg,cyl,wt) \%>\% lm_summary_tibble(mpg)
}
