% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibbleColumns.R
\name{prop_column_group}
\alias{prop_column_group}
\title{Proportion Column by Group}
\usage{
prop_column_group(df, group)
}
\description{
Groups one column, adds a column for count of each group 
and adds a column for proportion of total based on count
}
\examples{
mtcars \%>\% prop_column_group(cyl)
}
