% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibbleColumns.R
\name{replace_all_na}
\alias{replace_all_na}
\title{Replace all na's in data frame with a zero}
\usage{
replace_all_na(df)
}
\arguments{
\item{df}{a data frame}
}
\description{
Replace NA's with a zero for dataframes
}
\examples{
mtcars \%>\% tbl_out("cars") \%>\% sjmisc::set_na(na = 0) \%>\% replace_all_na()
}
