% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibbleColumns.R
\name{tbl_lookup}
\alias{tbl_lookup}
\title{Get a list of groups in certain column}
\usage{
tbl_lookup(df, ...)
}
\arguments{
\item{df, ...}{a data frame and a column or columns to get group list}
}
\description{
Designate a column and get a single column listing the groups in that column.
}
\examples{
mtcars \%>\% tbl_out("cars") \%>\% tbl_lookup(cyl) \%>\% tbl_out("cylList")
}
