% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibbleColumns.R
\name{tbl_module}
\alias{tbl_module}
\title{Run a function outside of the pipe sequence}
\usage{
tbl_module(df, fun, name)
}
\arguments{
\item{df, fun, name}{a data frame, a function to run and a name for created tibble object}
}
\description{
Run a function outside of the pipe sequence and create a tibble in global environment for it 
while passing previous state of data frame through to the next pipe step.
}
\examples{
mtcars \%>\% tbl_out("cars") \%>\% tbl_module(filter(.,hp > 150),"fastCars") \%>\% tbl_lookup(cyl) \%>\% tbl_out("cylList")
}
