% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibbleColumns.R
\name{tibble_out}
\alias{tibble_out}
\title{Tibble a data frame state within a pipe series}
\usage{
tibble_out(df, name)
}
\description{
Create a tibble for the state of a data frame within a pipe series
}
\examples{
mtcars \%>\% group_by(cyl) \%>\% prop_column_group(cyl) \%>\% tibble_out("grouped") \%>\% filter(Count >9)
}
