% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibbleColumns.R
\name{ttest_tibble}
\alias{ttest_tibble}
\title{A tidy t.test summary tibble}
\usage{
ttest_tibble(df1, df2)
}
\arguments{
\item{df1, df2}{two tibble dataframes}
}
\description{
Allows to pass a tibble data_frame to the base R t.test
function over two numeric columns. Then extracts the output
statistics and outputs a tibble.
}
\examples{
ttest_tibble(t1$num,t2$num)
}
