% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibbleColumns.R
\name{bool_to_binary}
\alias{bool_to_binary}
\title{New binary numeric column mirroring logical column}
\usage{
bool_to_binary(df, col, remove_bool_col = FALSE)
}
\arguments{
\item{df, col}{a data frame and logical TRUE FALSE column to change to binary}
}
\description{
Look at a logical TRUE FALSE column and make new column with binary representation.
}
\examples{
iris \%>\% mutate(Setosa = str_detect(.$Species, "setosa")) \%>\% bool_to_binary(.,Setosa)
}
