% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibbleColumns.R
\name{change_XoX_column}
\alias{change_XoX_column}
\title{General X over X change Column}
\usage{
change_XoX_column(df, col1, col2, XoX)
}
\arguments{
\item{df, col1, col2, XoX}{a data frame two columnnames and XoX name for new column}
}
\description{
Creates a change column based on integer or numeric column
}
\examples{
change_XoX_column(mtcars, drat, wt, "MoM")
}
