% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibbleColumns.R
\name{lm_summary_tibble}
\alias{lm_summary_tibble}
\title{Linear Model Summary Tibble}
\usage{
lm_summary_tibble(df, dep)
}
\description{
This function spits out a lm summary
}
\examples{
mtcars \%>\% prop_column_group(cyl)
}
