% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibbleColumns.R
\name{replace_all_na}
\alias{replace_all_na}
\title{Replace all na's in tibble with a value}
\usage{
replace_all_na(df, replace_with = NULL)
}
\arguments{
\item{df, replace_with}{a data frame and what to replace with}
}
\description{
Replace NA's with a zero for dataframes. Defaults with a 0 unless something else typed.
}
\examples{
mtcars \%>\% tbl_out("cars") \%>\% sjmisc::set_na(na = 0) \%>\% replace_all_na()
}
